chore: update generated client and test files
- Update auto-generated client code
- Add git sync integration test file
- Minor test file adjustments
